Clean up commented code

+++ src/main/java/com/exmek/core/persistence/repository/PlanetaryGearboxRepository.java
@@ -20,15 +20,6 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 
 	
 	//length
-//	@Query("SELECT MIN(p.length) FROM PlanetaryGearboxEntity p")
-//    Optional<BigDecimal> findMinLength();
-//
-//	@Query("SELECT MAX(p.length) FROM PlanetaryGearboxEntity p")
-//    Optional<BigDecimal> findMaxLength();
-//	
-//	@Query("SELECT DISTINCT p.lengthUnit FROM PlanetaryGearboxEntity p WHERE p.lengthUnit IS NOT NULL")
-//    Optional<List<String>> findLengthUnits();
-
 	@Query("""
 			SELECT MIN(m.length), MAX(m.length), m.lengthUnit 
 			FROM PlanetaryGearboxEntity m 
@@ -40,22 +31,12 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 	List<Object[]> findLengthMinMaxByUnits( 
 			@Param("ignoreSeries") int ignoreSeries, @Param("series") String series);
 	
-	default Map<LengthUnit, Range<Integer>> findLengthMinMaxByUnits(String series) {
+	default Map<LengthUnit, Range<BigDecimal>> findLengthMinMaxByUnits(String series) {
 		return JPAUtils.findMinMaxByUnits(series, this::findLengthMinMaxByUnits);
 	}
 	//
 
-
 	//weight
-//	@Query("SELECT MIN(p.weight) FROM PlanetaryGearboxEntity p")
-//    Optional<BigDecimal> findMinWeight();
-//
-//	@Query("SELECT MAX(p.weight) FROM PlanetaryGearboxEntity p")
-//    Optional<BigDecimal> findMaxWeight();
-//	
-//	@Query("SELECT DISTINCT p.weightUnit FROM PlanetaryGearboxEntity p WHERE p.weightUnit IS NOT NULL")
-//    Optional<List<String>> findWeightUnits();
-
 	@Query("""
 			SELECT MIN(m.weight), MAX(m.weight), m.weightUnit 
 			FROM PlanetaryGearboxEntity m 
@@ -72,17 +53,7 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 	}
 	//
 
-	
 	//frameSize
-//	@Query("SELECT MIN(p.frameSize) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMinFrameSize();
-//
-//	@Query("SELECT MAX(p.frameSize) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMaxFrameSize();
-//	
-//	@Query("SELECT DISTINCT p.frameSizeUnit FROM PlanetaryGearboxEntity p WHERE p.frameSizeUnit IS NOT NULL")
-//    Optional<List<String>> findFrameSizeUnits();
-
 	@Query("""
 			SELECT MIN(m.frameSize), MAX(m.frameSize), m.frameSizeUnit 
 			FROM PlanetaryGearboxEntity m 
@@ -99,14 +70,7 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 	}
 	//
 
-	
 	//nemaSize
-//	@Query("SELECT MIN(p.nemaSize) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMinNemaSize();
-//
-//	@Query("SELECT MAX(p.nemaSize) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMaxNemaSize();
-
 	@Query("""
 			SELECT MIN(m.nemaSize), MAX(m.nemaSize), '' 
 			FROM PlanetaryGearboxEntity m 
@@ -121,15 +85,8 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 		return JPAUtils.findMinMaxByUnits(series, this::findNemaSizeMinMaxByUnits);
 	}
 	//
-	
-	
-	//numOfStages
-//	@Query("SELECT MIN(p.numOfStages) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMinNumOfStages();
-//
-//	@Query("SELECT MAX(p.numOfStages) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMaxNumOfStages();
 
+	//numOfStages
 	@Query("""
 			SELECT MIN(m.numOfStages), MAX(m.numOfStages), '' 
 			FROM PlanetaryGearboxEntity m 
@@ -145,17 +102,7 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 	}
 	//
 
-
 	//efficiency
-//	@Query("SELECT MIN(p.efficiency) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMinEfficiency();
-//
-//	@Query("SELECT MAX(p.efficiency) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMaxEfficiency();
-//	
-//	@Query("SELECT DISTINCT p.efficiencyUnit FROM PlanetaryGearboxEntity p WHERE p.efficiencyUnit IS NOT NULL")
-//    Optional<List<String>> findEfficiencyUnits();
-
 	@Query("""
 			SELECT MIN(m.efficiency), MAX(m.efficiency), m.efficiencyUnit 
 			FROM PlanetaryGearboxEntity m 
@@ -172,17 +119,7 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 	}
 	//
 
-	
 	//ratedContinuousTorque
-//	@Query("SELECT MIN(p.ratedContinuousTorque) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMinRatedContinuousTorque();
-//
-//	@Query("SELECT MAX(p.ratedContinuousTorque) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMaxRatedContinuousTorque();
-//	
-//	@Query("SELECT DISTINCT p.ratedContinuousTorqueUnit FROM PlanetaryGearboxEntity p WHERE p.ratedContinuousTorqueUnit IS NOT NULL")
-//    Optional<List<String>> findRatedContinuousTorqueUnits();
-
 	@Query("""
 			SELECT MIN(m.ratedContinuousTorque), MAX(m.ratedContinuousTorque), m.ratedContinuousTorqueUnit 
 			FROM PlanetaryGearboxEntity m 
@@ -198,18 +135,8 @@ public interface PlanetaryGearboxRepository extends BaseProductRepository<Planet
 		return JPAUtils.findMinMaxByUnits(series, this::findRatedContinuousTorqueMinMaxByUnits);
 	}
 	//
-	
 
 	//maxMomentaryTorque
-//	@Query("SELECT MIN(p.maxMomentaryTorque) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMinMaxMomentaryTorque();
-//
-//	@Query("SELECT MAX(p.maxMomentaryTorque) FROM PlanetaryGearboxEntity p")
-//	Optional<BigDecimal> findMaxMaxMomentaryTorque();
-//	
-//	@Query("SELECT DISTINCT p.maxMomentaryTorqueUnit FROM PlanetaryGearboxEntity p WHERE p.maxMomentaryTorqueUnit IS NOT NULL")
-//    Optional<List<String>> findMaxMomentaryTorqueUnits();
-
 	@Query("""
 			SELECT MIN(m.maxMomentaryTorque), MAX(m.maxMomentaryTorque), m.maxMomentaryTorqueUnit 
 			FROM PlanetaryGearboxEntity m
